ingestion/pyproject.toml
[project]
name = "drbfm-assist-ingestion"
version = "0.1.0"
description = "DRBFM Assist Ingestion - Data ingestion job for Kubernetes"
requires-python = ">=3.13"
dependencies = [
    "loguru>=0.7.0",
    "python-dotenv>=1.0.0",
    "polars>=1.35.0",
    "pyyaml>=6.0.0",
    "trino>=0.329.0",
    "requests>=2.31.0",
    "httpx>=0.28.1",
    "authlib>=1.3.0",
    "pysocks>=1.7.1",
    "google-cloud-storage>=2.0.0",
    "google-cloud-bigquery>=3.38.0",
    "google-genai>=0.7.0",
    "google-cloud-aiplatform>=1.60.0",
    "google-auth>=2.0.0",
    "langfuse>=3.2.1",
    "click>=8.0.0",
    "pyarrow>=22.0.0",
    "google-cloud-bigquery-storage>=2.36.0",
    "psutil>=5.9.0",
    "langchain>=1.2.4",
    "langchain-text-splitters>=0.0.1",
]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = [
    "common",
    "common.gcs",
    "common.gemini",
    "common.isp",
    "common.m2m_token_issuer",
    "common.msqp",
    "common.pipelines",
    "common.pipelines.defects",
]

[tool.setuptools.package-data]
"*" = ["*.sh"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
    "pytest-cov>=6.0.0",
    "ruff>=0.14.9",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short -ra"
markers = [
    "unit: Unit tests (fast, isolated)",
    "integration: Integration tests (may be slower)",
    "slow: Slow tests",
]

[tool.coverage.run]
source = ["common"]
omit = ["tests/*", "*/tests/*", ".venv/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
